requirements.txt
asgiref==3.8.1
Django==5.0.6
django-crispy-forms==2.1
djangorestframework==3.15.1
sqlparse==0.5.0
